house_of_bloom/backend/requirements.txt
fastapi==0.114.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.26
psycopg2-binary==2.9.9
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.22
email-validator==2.0.0.post1
